@@ -13,7 +13,7 @@ impl From<TaikoApiError> for jsonrpsee_types::error::ErrorObject<'static> {
     fn from(error: TaikoApiError) -> Self {
         match error {
             TaikoApiError::GethNotFound => {
-                rpc_error_with_code(EthRpcErrorCode::ResourceNotFound.code(), format!("not found"))
+                rpc_error_with_code(EthRpcErrorCode::ResourceNotFound.code(), "not found".to_string())
             }
         }
     }
